refactor: split Database into one repository per aggregate
Replace the 1713-line database.rs — one Database impl exposing ~40
methods across six unrelated aggregates — with a database/ module
directory. Database shrinks to a connection holder with per-aggregate
accessors: db.templates(), db.settings(), db.recent_projects(),
db.team_libraries(), db.plugins().

- Each repository module owns its table's DDL, column migrations, row
  mappers, types, queries, and tests; Database::new runs each repo's
  init in FK order
- Sync log lives inside the team-libraries repository (FK-owned child,
  single consumer)
- Method names shorten within their repo (list_templates -> list,
  get_all_settings -> all, add_sync_log stays); bodies unchanged
- All public types re-exported from database/mod.rs, so
  crate::database::Template paths and the specta emit block are
  untouched; generated TS types are byte-identical
- Callers migrated mechanically: commands/*, team_library.rs, cli.rs
