+++ apps/desktop/src-tauri/src/commands/validation.rs
@@ -154,7 +154,7 @@ pub fn cmd_validate_schema(
     let loader = |name: &str| -> Option<TemplateData> {
         state_guard
             .db
-            .get_template_by_name(name)
+            .templates().get_by_name(name)
             .ok()
             .flatten()
             .map(|t| TemplateData {
